requirements0923.txt
absl-py==0.15.0
albumentations==0.4.6
argcomplete==1.8.1
astunparse==1.6.3
attrdict==2.0.1
attrs==20.3.0
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
black==22.8.0
blinker==1.4
bottle==0.12.19
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cppman==0.5.0
cryptography==3.3.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
distlib==0.3.4
distro==1.5.0
distro-info==1.0
docker==4.1.0
docutils==0.16
filelock==3.6.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
Glances==3.1.5
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==0.4.17
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.10
imageio==2.16.1
imgaug==0.4.0
importlib-metadata==4.11.3
incremental==17.5.0
influxdb==5.2.3
iniconfig==1.1.1
jeepney==0.6.0
Jinja2==2.11.2
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==22.2.0
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lxml==4.6.3
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.3.4
more-itertools==4.2.0
msgpack==1.0.0
mypy-extensions==0.4.3
netaddr==0.7.19
netifaces==0.10.9
networkx==2.7.1
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
onnx==1.11.0
onnxoptimizer==0.2.6
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pathspec==0.10.1
pbr==5.8.1
pexpect==4.8.0
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
protobuf==3.20.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pycryptodomex==3.14.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5b0
pysmi==0.3.2
pysnmp==4.4.12
pystache==0.5.4
pytest==7.1.0
python-dateutil==2.8.1
python-debian==0.1.39
pytz==2021.1
PyWavelets==1.3.0
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.1
roman==2.0.0
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
SCons==4.0.1
SecretStorage==3.3.1
service-identity==18.1.0
Shapely==1.8.1.post1
simplejson==3.17.2
six==1.15.0
soupsieve==2.3.2.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.2
tensorflow-estimator==2.5.0
termcolor==1.1.0
testresources==2.0.1
tflite==2.1.0
tifffile==2022.4.28
toml==0.10.2
tomli==2.0.1
torch==1.7.0
torchvision==0.8.1
tornado==6.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
